@@ -4,17 +4,17 @@ version: 0.4.0
 publish_to: none
 
 environment:
-  sdk: '>=3.0.1 <4.0.0'
+  sdk: '>=3.1.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 #  image_picker_web: ^3.1.1
   # state management
-  provider: ^6.0.4
+  provider: ^6.1.2
   # image
-  image: ^3.2.2
-  path_provider: ^2.0.11
+  image: ^4.1.7
+  path_provider: ^2.1.2
   photo_view: ^0.14.0
   image_gallery_saver: ^2.0.3
   #own packages
@@ -26,20 +26,20 @@ dependencies:
 
   modal_gif_picker: ^0.0.2
   # painting
-  perfect_freehand: ^1.0.4
+  perfect_freehand: ^2.3.2
   # positioning
-  align_positioned: ^3.0.0
+  align_positioned: ^4.0.1
   #message
-  fluttertoast: ^8.1.1
+  fluttertoast: ^8.2.4
 
   animated_text_kit: ^4.2.2
 
-  flutter_screenutil: ^5.8.2
+  flutter_screenutil: ^5.9.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^3.0.1
 
 
 flutter:
