@@ -11,7 +11,7 @@ class Sketcher extends CustomPainter {
   @override
   void paint(Canvas canvas, Size size) {
     Paint paint = Paint();
-    List<Point>? outlinePoints;
+    List<Offset>? outlinePoints;
 
     for (int i = 0; i < lines.length; ++i) {
       switch (lines[i].paintingType) {
@@ -19,10 +19,9 @@ class Sketcher extends CustomPainter {
           paint = Paint()..color = lines[i].lineColor;
 
           outlinePoints = getStroke(
-
-              /// coordinates
-              lines[i].points,
-
+            /// coordinates
+            lines[i].points,
+            options: StrokeOptions(
               /// line width
               size: lines[i].size,
 
@@ -35,11 +34,9 @@ class Sketcher extends CustomPainter {
               /// on complete line
               isComplete: lines[i].isComplete,
               streamline: 1,
-              taperEnd: 0,
-              taperStart: 0,
-              capEnd: true,
               simulatePressure: true,
-              capStart: true);
+            ),
+          );
           break;
         case PaintingType.marker:
           paint = Paint()
@@ -52,18 +49,19 @@ class Sketcher extends CustomPainter {
           outlinePoints = getStroke(
             /// coordinates
             lines[i].points,
+            options: StrokeOptions(
+              /// line width
+              size: lines[i].size,
 
-            /// line width
-            size: lines[i].size,
+              /// line thin
+              thinning: 1,
 
-            /// line thin
-            thinning: 1,
+              /// line smooth
+              smoothing: 1,
 
-            /// line smooth
-            smoothing: 1,
-
-            /// on complete line
-            isComplete: lines[i].isComplete,
+              /// on complete line
+              isComplete: lines[i].isComplete,
+            ),
           );
           break;
         case PaintingType.neon:
@@ -78,10 +76,9 @@ class Sketcher extends CustomPainter {
             ..style = PaintingStyle.stroke;
 
           outlinePoints = getStroke(
-
-              /// coordinates
-              lines[i].points,
-
+            /// coordinates
+            lines[i].points,
+            options: StrokeOptions(
               /// line width
               size: lines[i].size,
 
@@ -95,10 +92,8 @@ class Sketcher extends CustomPainter {
               isComplete: lines[i].isComplete,
               streamline: lines[i].streamline,
               simulatePressure: lines[i].simulatePressure,
-              taperStart: 0,
-              taperEnd: 0,
-              capStart: true,
-              capEnd: true);
+            ),
+          );
           break;
       }
 
@@ -109,16 +104,17 @@ class Sketcher extends CustomPainter {
       } else if (outlinePoints.length < 2) {
         /// If the path only has one line, draw a dot.
         path.addOval(Rect.fromCircle(
-            center: Offset(outlinePoints[0].x, outlinePoints[0].y), radius: 1));
+            center: Offset(outlinePoints[0].dx, outlinePoints[0].dy),
+            radius: 1));
       } else {
         /// Otherwise, draw a line that connects each point with a curve.
-        path.moveTo(outlinePoints[0].x, outlinePoints[0].y);
+        path.moveTo(outlinePoints[0].dx, outlinePoints[0].dy);
 
         for (int i = 1; i < outlinePoints.length - 1; ++i) {
           final p0 = outlinePoints[i];
           final p1 = outlinePoints[i + 1];
           path.quadraticBezierTo(
-              p0.x, p0.y, (p0.x + p1.x) / 2, (p0.y + p1.y) / 2);
+              p0.dx, p0.dy, (p0.dx + p1.dx) / 2, (p0.dy + p1.dy) / 2);
         }
       }
 
