@@ -34,7 +34,7 @@ class ColorDetection {
     double py = localPosition.dy;
 
     // int pixel32 = photo!.getPixelIndex(px.toInt(), py.toInt());
-    int pixel32 = photo!.getPixel(px.toInt(), py.toInt());
+    int pixel32 = photo!.getPixelIndex(px.toInt(), py.toInt());
 
     int hex = abgrToArgb(pixel32);
 
@@ -43,9 +43,11 @@ class ColorDetection {
   }
 
   Future<void> loadSnapshotBytes() async {
-    RenderRepaintBoundary? boxPaint = paintKey!.currentContext!.findRenderObject() as RenderRepaintBoundary?;
+    RenderRepaintBoundary? boxPaint =
+        paintKey!.currentContext!.findRenderObject() as RenderRepaintBoundary?;
     ui.Image capture = await boxPaint!.toImage();
-    ByteData? imageBytes = await capture.toByteData(format: ui.ImageByteFormat.png);
+    ByteData? imageBytes =
+        await capture.toByteData(format: ui.ImageByteFormat.png);
     setImageBytes(imageBytes!);
     capture.dispose();
   }
